+++ mbo/diff/unified_diff.h
@@ -46,11 +46,7 @@ class UnifiedDiff final {
   UnifiedDiff() = delete;
 
   struct Options final {
-    static const Options& Default() noexcept {
-      // This function can be `constexpr` with GCC 12
-      const types::NoDestruct<UnifiedDiff::Options> defaults{};
-      return defaults.Get();
-    }
+    static const Options& Default() noexcept;
 
     std::size_t context_size = 3;
     std::string time_format = "%F %H:%M:%E3S %z";